added homebrew as a alternative dependency source

+++ package_osx.sh
@@ -1,6 +1,7 @@
 #!/usr/bin/env bash -e
 
 FLARE_EXE=$1
+FLARE_DEPS_SRC="http"
 
 if [ -z "$FLARE_EXE" ]; then
   echo "usage: $0 <path to flare executable>"
@@ -15,6 +16,19 @@ if [ `otool -L ${FLARE_EXE} | egrep libSDL2 | wc -l` -lt 1 ]; then
   exit 1
 fi
 
+if [ "$2" != "" ]; then
+ FLARE_DEPS_SRC=$2
+fi
+if [ $FLARE_DEPS_SRC == "http" ]; then
+  echo "download dependencies from website"
+elif [ $FLARE_DEPS_SRC == "homebrew" ]; then
+  echo "copy dependencies from homebrew"
+else
+  echo "usage: $0 <path to flare executable> <http|homebrew>"
+  exit 1
+fi
+
+
 DST=/tmp/___flare.build
 rm -fr ${DST} && mkdir -p ${DST}
 
@@ -25,10 +39,30 @@ cp -r RELEASE_NOTES.txt \
   ${FLARE_EXE} \
   mods ${DST}
 
-#feel free to build dependencies by yourself btw
-wget 'http://files.ruads.org/flare_osx_dependencies.tar.gz' -P ${DST}
-tar -zxf ${DST}/flare_osx_dependencies.tar.gz -C ${DST}
-rm -f ${DST}/flare_osx_dependencies.tar.gz 
+if [ $FLARE_DEPS_SRC == "http" ]; then
+  #feel free to build dependencies by yourself btw
+  wget 'http://files.ruads.org/flare_osx_dependencies.tar.gz' -P ${DST}
+  tar -zxf ${DST}/flare_osx_dependencies.tar.gz -C ${DST}
+  rm -f ${DST}/flare_osx_dependencies.tar.gz
+elif [ $FLARE_DEPS_SRC == "homebrew" ]; then
+  LIB=${DST}/lib
+  mkdir $LIB
+  # SDL2
+  cp /usr/local/opt/sdl2/COPYING.txt $LIB/SDL2-COPYING.txt
+  cp /usr/local/opt/sdl2/README.txt $LIB/SDL2-README.txt
+  cp /usr/local/opt/sdl2/lib/libSDL2-2.0.0.dylib $LIB
+  cp /usr/local/opt/sdl2_image/lib/libSDL2_image-2.0.0.dylib $LIB
+  cp /usr/local/opt/sdl2_mixer/lib/libSDL2_mixer-2.0.0.dylib $LIB
+  cp /usr/local/opt/sdl2_ttf/lib/libSDL2_ttf-2.0.0.dylib $LIB
+  # VORBIS
+  cp /usr/local/opt/libvorbis/lib/COPYING $LIB/VORBIS-COPYING
+  cp /usr/local/opt/libvorbis/lib/libvorbisfile.0.dylib $LIB
+  cp /usr/local/opt/libvorbis/lib/libvorbisenc.2.dylib $LIB
+  cp /usr/local/opt/libvorbis/lib/libvorbisfile.3.dylib $LIB
+else
+  echo "'$FLARE_DEPS_SRC' unknown dependency source"
+  exit 1
+fi
 
 echo '#!/bin/sh' >> ${DST}/start.sh
 echo 'cd "$(dirname "${BASH_SOURCE[0]}")"' >> ${DST}/start.sh
@@ -37,5 +71,5 @@ chmod +x ${DST}/start.sh
 
 echo "packaging"
 tar -zcf flare_osx.tar.gz -C ${DST} .
-rm -fr ${DST}
+#rm -fr ${DST}
 echo "done"
